+++ src/main/java/com/raoulvdberge/refinedstorage/block/BlockSolderer.java
@@ -10,11 +10,15 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
+import net.minecraft.util.EnumParticleTypes;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
 import java.util.List;
+import java.util.Random;
 
 public class BlockSolderer extends BlockNode {
     public BlockSolderer() {
@@ -41,6 +45,48 @@ public class BlockSolderer extends BlockNode {
             "block.refinedstorage:solderer.tooltip.2",
             TextFormatting.WHITE + I18n.format("block.refinedstorage:cable.name") + TextFormatting.GRAY
         ));
+    }
+
+    @Override
+    @SideOnly(Side.CLIENT)
+    public void randomDisplayTick(IBlockState state, World world, BlockPos pos, Random rand) {
+        TileEntity tile = world.getTileEntity(pos);
+
+        if (tile instanceof TileSolderer && ((TileSolderer) tile).isWorking()) {
+            EnumFacing direction = getActualState(state, world, pos).getValue(DIRECTION);
+
+            double x = 0;
+            double y = (double) pos.getY() + 0.6D + rand.nextDouble() / 32F;
+            double z = 0;
+
+            if (direction == EnumFacing.NORTH) {
+                x = (double) pos.getX() + 0.4D;
+                z = (double) pos.getZ() + 0.4D;
+            } else if (direction == EnumFacing.EAST) {
+                x = (double) pos.getX() + 0.6D;
+                z = (double) pos.getZ() + 0.4D;
+            } else if (direction == EnumFacing.SOUTH) {
+                x = (double) pos.getX() + 0.6D;
+                z = (double) pos.getZ() + 0.6D;
+            } else if (direction == EnumFacing.WEST) {
+                x = (double) pos.getX() + 0.4D;
+                z = (double) pos.getZ() + 0.6D;
+            }
+
+            int particles = rand.nextInt(5);
+
+            for (int i = 0; i < 1 + particles; ++i) {
+                world.spawnParticle(
+                    EnumParticleTypes.SMOKE_NORMAL,
+                    x + (rand.nextDouble() / 16F * (rand.nextBoolean() ? 1 : -1)),
+                    y,
+                    z + (rand.nextDouble() / 16F * (rand.nextBoolean() ? 1 : -1)),
+                    0.0D,
+                    0.0D,
+                    0.0D
+                );
+            }
+        }
     }
 
     @Override